Prepare to build vRouter as Hyper-V Extension
Changes required to support Contrail on Windows.
1. Add OS-specific GUID to vr_interface
2. Use compiler-independent wrappers on __attribute__((packed))
3. Use compiler-independent wrappers on __attribute__((format))
4. Use compiler-independent wrappers on __attribute__((unused))
5. Cast pointers to uintptr_t instead of unsigned long

Partial-Bug: #1734699

+++ dp-core/vr_nexthop.c
@@ -1037,7 +1037,7 @@ nh_handle_unknown_unicast(struct vr_packet *pkt, struct vr_eth *eth,
         return !handled;
 
     /* Get the packet to eth header */
-    pull_len = (unsigned long)eth - (unsigned long)pkt_data(pkt);
+    pull_len = (uintptr_t)eth - (uintptr_t)pkt_data(pkt);
     if (pull_len)
         pkt_pull(pkt, pull_len);
